+++ src/syntax/abs/desugar/decls.rs
@@ -3,7 +3,7 @@ use crate::syntax::surf::ExprDecl;
 use super::{DesugarErr, DesugarM, DesugarState};
 
 pub fn desugar_decls(state: DesugarState, decls: Vec<ExprDecl>) -> DesugarM {
-    decls.into_iter().try_fold(tcs, desugar_decl)
+    decls.into_iter().try_fold(state, desugar_decl)
 }
 
 pub fn desugar_decl(state: DesugarState, decl: ExprDecl) -> DesugarM {