+++ src/syntax/abs/desugar/error.rs
@@ -2,7 +2,6 @@ use std::fmt::{Display, Error as FmtError, Formatter};
 
 use voile_util::loc::Ident;
 
-#[derive(Debug, Clone)]
 pub enum DesugarErr {
     UnresolvedReference(Ident),
 }