+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 + Unfolds data and codata applications (#27)
 + Parsing (expr parsing and file parsing) (#20, #31)
 + `Abs::App` is now chained (#33)
++ Desugar surface into abstract, scope check (#36)
 
 # 0.0.3